Provide a TermDomain instance for AnyDomain
This is a bit hacky and may not always work as expected as all
"constructing" methods will produce a TopDomain whereas all "consuming"
methods will expect the passed AnyDomains to be of the same domain type.
This means it is unfortunately easy to accidentally introduce
heterogeneity when not desired, but for now it should hopefully be
enough to at least let us prettyprint the types.